seed user, categories, product

+++ migrations/20231030205921-create-order.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       total_price: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL
       },
       AddressId: {
         type: Sequelize.INTEGER
